Move markdown loading into its own module and convert to TypeScript
